readBytes needed to return -1 on EOS. Not sure how this will work on pipes. But it's how IKVM used to do it.

+++ src/IKVM.Runtime/Java/Externs/sun/nio/ch/SocketDispatcher.cs
@@ -40,7 +40,7 @@ namespace IKVM.Java.Externs.sun.nio.ch
         /// <param name="address"></param>
         /// <param name="len"></param>
         /// <returns></returns>
-        public static unsafe long read0(object fd, long address, int len)
+        public static unsafe int read0(object fd, long address, int len)
         {
 #if FIRST_PASS
             throw new NotImplementedException();
